+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Change Log
 
 ### master (untagged)
 
+* Add preload and postload
+
 ### 0.0.0 (2017-02-09)
 
 * Initial release